pyproject.toml
[project]
name = "moviellms"
version = "0.1.0"
description = "Usage of LLMs and LangChain on dataset with movie descriptions"
authors = [
    {name = "Jedrek00",email = "jedrek.koscielniak@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain (>=0.3.18,<0.4.0)",
    "langchain-community (>=0.3.17,<0.4.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchain-openai (>=0.3.6,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.6"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "N",  # PEP8 naming convetions
    "D"  # pydocstyle
]

[tool.ruff.lint.pydocstyle]
convention = "google"